requirements.txt
# OmniTrader MVP Dependencies
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
pyyaml>=6.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-dotenv>=1.0.0
structlog>=23.0.0
